removed getTimerState function from buttion handle functions in new components
