+++ raynian-fe/src/components/timer/TimerContols.jsx
@@ -11,41 +11,31 @@ const TimerContols = () => {
   const timerState = useSelector((state) => state.timer);
   const { roomId } = useSelector((state) => state.room);
 
-  const getTimerState = async () => {
-    const timerData = await dispatch(timerActions.getTimerState());
-    if (roomId != null) {
-      const updatedTimerData = {
-        countdown: timerData.countdown,
-        isRunning: timerData.isRunning,
-        isBreak: timerData.isBreak,
-        isWork: timerData.isWork,
-        isPaused: timerData.isPaused,
-        workTime: timerData.workTime,
-        breakTime: timerData.breakTime,
-      };
+  // const getTimerState = async () => {
+  //   const timerData = await dispatch(timerActions.getTimerState());
+  //   if (roomId != null) {
+  //     const updatedTimerData = {
+  //       countdown: timerData.countdown,
+  //       isRunning: timerData.isRunning,
+  //       isBreak: timerData.isBreak,
+  //       isWork: timerData.isWork,
+  //       isPaused: timerData.isPaused,
+  //       workTime: timerData.workTime,
+  //       breakTime: timerData.breakTime,
+  //     };
 
-      updateTimerStatus(updatedTimerData, roomId);
-    }
-  };
-
-  const handleWorkTimeChange = (e) => {
-    dispatch(timerActions.setWorkTime(parseInt(e.target.value)));
-    getTimerState();
-  };
-
-  const handleBreakTimeChange = (e) => {
-    dispatch(timerActions.setBreakTime(parseInt(e.target.value)));
-    getTimerState();
-  };
+  //     updateTimerStatus(updatedTimerData, roomId);
+  //   }
+  // };
 
   const handleStartTimer = () => {
     dispatch(timerActions.startTimer());
-    getTimerState();
+    // getTimerState();
   };
 
   const handleStopTimer = () => {
     dispatch(timerActions.stopTimer());
-    getTimerState();
+    // getTimerState();
   };
 
   const handlePauseTimer = () => {
@@ -55,29 +45,6 @@ const TimerContols = () => {
 
   return (
     <div className="flex flex-col items-start">
-      <div>
-        <label htmlFor="workTime">Work Time:</label>
-        <select
-          id="workTime"
-          value={timerState.workTime}
-          onChange={handleWorkTimeChange}
-        >
-          <option value={60}>60 minutes</option>
-          <option value={45}>45 minutes</option>
-          <option value={1}>30 minutes</option>
-        </select>
-
-        <label htmlFor="breakTime">Break Time:</label>
-        <select
-          id="breakTime"
-          value={timerState.breakTime}
-          onChange={handleBreakTimeChange}
-        >
-          <option value={15}>15 minutes</option>
-          <option value={10}>10 minutes</option>
-          <option value={1}>5 minutes</option>
-        </select>
-      </div>
       <button onClick={handleStartTimer}>StartTimer</button>
       <button onClick={handleStopTimer}>StopTimer</button>
       <button onClick={handlePauseTimer}>PauseTimer</button>
